Move options structs from neo_skeleton to options
