+++ components/datetime/src/fieldset.rs
@@ -10,7 +10,7 @@ use crate::{
     dynamic::*,
     fields,
     format::neo::*,
-    neo_skeleton::*,
+    options::*,
     provider::{neo::*, time_zones::tz, *},
     raw::neo::RawNeoOptions,
     scaffold::*,
