+++ components/datetime/tests/fixtures/mod.rs
@@ -4,7 +4,8 @@
 
 #![cfg(feature = "serde")]
 
-use icu_datetime::{neo_skeleton, provider::skeleton::components};
+use icu_datetime::options;
+use icu_datetime::provider::skeleton::components;
 use icu_locale_core::preferences::extensions::unicode::keywords::HourCycle;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -64,7 +65,7 @@ pub struct TestComponentsBag {
     pub hour: Option<components::Numeric>,
     pub minute: Option<components::Numeric>,
     pub second: Option<components::Numeric>,
-    pub fractional_second: Option<neo_skeleton::FractionalSecondDigits>,
+    pub fractional_second: Option<options::FractionalSecondDigits>,
 
     pub time_zone_name: Option<components::TimeZoneName>,
 }
