+++ tutorials/rust/buffer/src/main.rs
@@ -14,7 +14,7 @@
 //! For more information, see the tutorial [cargo.md](../../cargo.md).
 
 use icu::calendar::{DateTime, Gregorian};
-use icu::datetime::{FixedCalendarDateTimeFormatter, fieldset::YMDT, neo_skeleton::NeoSkeletonLength};
+use icu::datetime::{FixedCalendarDateTimeFormatter, fieldset::YMDT, options::NeoSkeletonLength};
 use icu::locale::locale;
 use icu_provider_blob::BlobDataProvider;
